# projects/ecommerce_store/backend/backend_app/api/routes/product.py
# GENERATED by the Backend Agent from the project graphs. Do not edit by hand: regenerate instead.
"""HTTP routes for service.product. Route handlers only delegate to the application service."""
from __future__ import annotations

from decimal import Decimal
from typing import Literal
from uuid import UUID

from fastapi import APIRouter, Depends, Query, Response

from ...api_common import error_responses
from ...auth import authorize
from ...engine import ListQuery
from ..deps import get_services
from ...application.services import Services
from ..schemas.product import ProductResponse


router = APIRouter(tags=["product"])


@router.get(
    "/products",
    operation_id="api.product.list",
    summary="List Products",
    description="List Products.\n\nOperation `operation.product.list`. Access: public.",
    response_model=list[ProductResponse],
    status_code=200,
    responses=error_responses(authenticated=False, restricted=False, has_path_id=False, mutating=False, rate_limited=False),
)
def api_product_list(
    response: Response,
    limit: int = Query(100, ge=1, le=1000),
    offset: int = Query(0, ge=0),
    sort: str | None = Query(None, max_length=64, description="Attribute to sort by"),
    order: Literal['asc', 'desc'] = Query('asc'),
    principal: None = Depends(authorize("operation.product.list")),
    services: Services = Depends(get_services),
):
    page = services.product.list(principal, ListQuery(filters={}, search=None, sort=sort, order=order, limit=limit, offset=offset))
    response.headers['X-Total-Count'] = str(page.total)
    return page.items


@router.get(
    "/products/filter",
    operation_id="api.product.filter",
    summary="Filter products",
    description="Filter products.\n\nOperation `operation.product.filter`. Access: public.",
    response_model=list[ProductResponse],
    status_code=200,
    responses=error_responses(authenticated=False, restricted=False, has_path_id=False, mutating=False, rate_limited=False),
)
def api_product_filter(
    response: Response,
    category_id: UUID | None = None,
    price: Decimal | None = None,
    limit: int = Query(100, ge=1, le=1000),
    offset: int = Query(0, ge=0),
    sort: str | None = Query(None, max_length=64, description="Attribute to sort by"),
    order: Literal['asc', 'desc'] = Query('asc'),
    principal: None = Depends(authorize("operation.product.filter")),
    services: Services = Depends(get_services),
):
    page = services.product.filter(principal, ListQuery(filters={'category_id': category_id, 'price': price}, search=None, sort=sort, order=order, limit=limit, offset=offset))
    response.headers['X-Total-Count'] = str(page.total)
    return page.items


@router.get(
    "/products/search",
    operation_id="api.product.search",
    summary="Search products",
    description="Search products.\n\nOperation `operation.product.search`. Access: public.",
    response_model=list[ProductResponse],
    status_code=200,
    responses=error_responses(authenticated=False, restricted=False, has_path_id=False, mutating=False, rate_limited=False),
)
def api_product_search(
    response: Response,
    query: str | None = Query(None, max_length=200, description="Case-insensitive text search"),
    limit: int = Query(100, ge=1, le=1000),
    offset: int = Query(0, ge=0),
    sort: str | None = Query(None, max_length=64, description="Attribute to sort by"),
    order: Literal['asc', 'desc'] = Query('asc'),
    principal: None = Depends(authorize("operation.product.search")),
    services: Services = Depends(get_services),
):
    page = services.product.search(principal, ListQuery(filters={}, search=query, sort=sort, order=order, limit=limit, offset=offset))
    response.headers['X-Total-Count'] = str(page.total)
    return page.items


@router.get(
    "/products/{id}",
    operation_id="api.product.read",
    summary="View Product",
    description="View Product.\n\nOperation `operation.product.read`. Access: public.",
    response_model=ProductResponse,
    status_code=200,
    responses=error_responses(authenticated=False, restricted=False, has_path_id=True, mutating=False, rate_limited=False),
)
def api_product_read(
    id: UUID,
    principal: None = Depends(authorize("operation.product.read")),
    services: Services = Depends(get_services),
):
    return services.product.read(principal, id)