# projects/ecommerce_store/backend/backend_app/api/schemas/product.py
# GENERATED by the Backend Agent from the project graphs. Do not edit by hand: regenerate instead.
"""API request/response models for entity.product (from api.json schemas)."""
from __future__ import annotations

from uuid import UUID

from pydantic import BaseModel, ConfigDict

from ...api_common import JsonDecimal


class ProductResponse(BaseModel):
    """schema.product: Product"""

    model_config = ConfigDict(from_attributes=True)
    id: UUID
    name: str
    description: str | None = None
    price: JsonDecimal
    stock: int
    category_id: UUID | None = None
    image_url: str | None = None